Unauthorized response now 401 error, not redirect
The original response when someone was not authorized to view a url was
to return a redirect to the login page. This made it difficult for a
client to decide what to do with the response, since it ended up with a
200 OK, but the contents were a login page. This change turns all
the unauthorized requests into a 401 error, which is easy to trap and
then use to require the client to login again.