Adds failure if response status is not 200.
Without this check the code would proceed to try and get the user
profile information and fail. The failure was causing the code to crash,
but the server was catching the crash and restarting. The catch and
restart did not working when running the server through a debugger and
caused the process to terminate.